+++ src/mastercore_tx.cpp
@@ -121,14 +121,14 @@ int CMPTransaction::step2_Alert(std::string *new_global_alert_message)
           int32_t alertType;
           uint64_t expiryValue;
           uint32_t typeCheck;
-          uint8_t verCheck;
+          uint32_t verCheck;
           string alertMessage;
           try
           {
               alertType = boost::lexical_cast<int32_t>(vstr[0]);
               expiryValue = boost::lexical_cast<uint64_t>(vstr[1]);
               typeCheck = boost::lexical_cast<uint32_t>(vstr[2]);
-              verCheck = boost::lexical_cast<uint8_t>(vstr[3]);
+              verCheck = boost::lexical_cast<uint32_t>(vstr[3]);
           } catch (const boost::bad_lexical_cast &e)
             {
                   file_log("DEBUG ALERT - error in converting values from global alert string\n");
